IRGen: Move linkage computation to Linkage.cpp
